### 2022-03-04 #505 @frankfarach
npi: Access the U.S. National Provider Identifier Registry API
**Reviewers:**
Submitting Author Name: Frank Farach
Submitting Author Github Handle: @frankfarach
Repository:  https://github.com/frankfarach/npi
Version submitted: 0.1.0
Submission type: Standard

Version accepted: TBD
---

```
Package: npi
Title: Access the U.S. National Provider Identifier Registry API
Version: 0.1.0
    person("Frank", "Farach", , "frank.farach@gmail.com", role = c("cre", "aut"),
    person("Sam", "Parmar", , "parmartsam@gmail.com", role = "ctb")
  )
Description: Access the United States National Provider Identifier
    Registry API (if available) and provide informative error messages
    when it's not.
    https://frankfarach.github.io/npi/,
    https://npiregistry.cms.hhs.gov/api/
    checkLuhn,
    glue,
    httr,
    utils
    checkmate,
    covr,
    httptest,
    mockery,
    knitr
```

npi is an API wrapper for accessing and retrieving data from the public U.S. National Provider Identifier Registry.

Healthcare analysts, public health analysts, health economists, health policy analysts, medical informaticists, and journalists can use data obtained through this package to address a wide range of scientific questions in their respective domains. NPI numbers appear in other publicly available records, making them especially useful for linking datasets (e.g., healthcare quality, Medicare, etc.) in an otherwise fragmented healthcare data landscape.

None that I could find on Metacran when searching for [npi](https://www.r-pkg.org/search.html?q=npi) or [NPPES](https://www.r-pkg.org/search.html?q=nppes).

Yes. Although the API, and thus the package, returns personally identifiable information (e.g., names of individual providers), the U.S. government already makes these records freely and publicly available [online](https://npiregistry.cms.hhs.gov/), and provides [data dissemination notices](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPPES_FOIA_Data-Elements_062007.pdf) on the data elements to be disclosed from NPI records.

The only failing check is for continuous integration. However, the package repo performs CI integration through Github Actions, including standard R-CMD-check, test coverage, pkgdown, and CITATION.cff.
