+++ vignettes/advanced-use.Rmd
@@ -1,0 +1,58 @@
+---
+title: "Advanced Uses"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{advanced-use}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(npi)
+library(purrr)
+```
+
+This vignette provides an brief exploration to advanced uses of the npi package.
+
+## Run `npi_search()` on multiple search terms
+
+Users can serially execute searches for multiple values of a single query parameter by using `npi` in combination with the [`purrr`](https://purrr.tidyverse.org/) package. In the example below, we search multiple NPI numbers. A single tibble is returned with record information corresponding toe matching records.
+
+```{r}
+npis <- c(1992708929, 1831192848, 1699778688, 1111111111)  # Last element doesn't exist
+
+out <- npis %>% 
+  purrr::map(., ~ npi_search(number = .)) %>% 
+  dplyr::bind_rows()
+
+npi_summarize(out)
+```
+
+Here we search for multiple zip codes in Los Angeles County.
+
+```{r}
+codes <- c(90210, 90211, 90212)
+
+zip_3 <- codes %>% 
+  purrr::map(., ~ npi_search(postal_code  = .)) %>% 
+  dplyr::bind_rows() 
+
+npi_flatten(zip_3)
+```
+
+Consult the R for Data Science [chapter on iteration](https://r4ds.had.co.nz/iteration.html) to learn more about using the `purrr` package.
+
+## Note on NPI Downloadable Files
+
+CMS regularly releases full NPI data files [here](https://download.cms.gov/nppes/NPI_Files.html). Consider downloading the entire data set as needed. Data dissemination files are zipped and will exceed 4GB upon decompression. Additional documentation is provided [here](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/DataDissemination).
+
+
+
+
