+++ vignettes/advanced-use.Rmd
@@ -19,7 +19,21 @@ library(npi)
 library(purrr)
 ```
 
-This vignette provides an brief exploration to advanced uses of the npi package.
+This vignette explores advanced uses of the npi package.
+
+`npi` is an R package that allows R users to access the [U.S. National Provider Identifier (NPI) Registry](https://npiregistry.cms.hhs.gov/) API by the Center for Medicare and Medicaid Services (CMS). 
+
+The package makes it easy to obtain administrative data linked to a specific individual or organizational healthcare provider. Additionally, users can perform advanced searches based on provider name, location, type of service, credentials, and many other attributes.
+
+See the npi::npi vignette for an introduction to the package.
+
+List of topics covered in this vignette:
+* Note on NPI Downloadable Files
+* Run `npi_search()` on multiple search terms
+
+## Note on NPI Downloadable Files
+
+CMS regularly releases full NPI data files [here](https://download.cms.gov/nppes/NPI_Files.html). We recommend that users download the data file if they need to work with the entire dataset. The API and `npi_search()` returns a maximum of 1,200 records. Also consider downloading the entire data if you need to work with more than the maximum. Data dissemination files are zipped and will exceed 4GB upon decompression. Additional documentation is provided [here](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/DataDissemination).
 
 ## Run `npi_search()` on multiple search terms
 
@@ -48,10 +62,6 @@ npi_flatten(zip_3)
 ```
 
 Consult the R for Data Science [chapter on iteration](https://r4ds.had.co.nz/iteration.html) to learn more about using the `purrr` package.
-
-## Note on NPI Downloadable Files
-
-CMS regularly releases full NPI data files [here](https://download.cms.gov/nppes/NPI_Files.html). Consider downloading the entire data set as needed. Data dissemination files are zipped and will exceed 4GB upon decompression. Additional documentation is provided [here](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/DataDissemination).
 
 
 
