### 2022-03-12 bd22fb1
added toc to both vignettes
