+++ vignettes/advanced-use.Rmd
@@ -33,7 +33,7 @@ CMS regularly releases full NPI data files [here](https://download.cms.gov/nppes
 
 ## Run `npi_search()` on multiple search terms
 
-Users can serially execute searches for multiple values of a single query parameter by using `npi` in combination with the [`purrr`](https://purrr.tidyverse.org/) package. In the example below, we search multiple NPI numbers. A single tibble is returned with record information corresponding toe matching records.
+Users can serially execute searches for multiple values of a single query parameter by using `npi` in combination with the [`purrr`](https://purrr.tidyverse.org/) package. In the example below, we search multiple NPI numbers. A single tibble is returned with record information corresponding toe matching records.The [purrr:map()](https://purrr.tidyverse.org/reference/map.html) function is used to apply the `npi_search()` function on each element of the vector. Therafter, the [dplyr::bind_rows()](https://dplyr.tidyverse.org/reference/bind.html) function is used to combine the list of dataframes together into a single dataframe.
 
 ```{r}
 npis <- c(1992708929, 1831192848, 1699778688, 1111111111)  # Last element doesn't exist
@@ -58,6 +58,19 @@ npi_flatten(zip_3)
 ```
 
 Consult the R for Data Science [chapter on iteration](https://r4ds.had.co.nz/iteration.html) to learn more about using the `purrr` package.
+
+Alternatively, you can use a simple for loop instead if you are unfamiliar with the tidyverse approach.
+
+```{r}
+npis <- c(1992708929, 1831192848, 1699778688, 1111111111)  # Last element doesn't exist
+combined_df  <- data.frame()
+for(i in npis){
+  combined_df <- rbind(combined_df, npi_search(number = i))
+}
+
+npi_summarize(combined_df)
+```
+
 
 
 
