+++ man/npi-package.Rd
@@ -8,7 +8,9 @@
 \description{
 \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
-Access the United States National Provider Identifier Registry API (if available) and provide informative error messages when it's not.
+Access the United States National Provider Identifier
+    Registry API (if available) and provide informative error messages
+    when it's not.
 }
 \details{
 \pkg{npi} makes it easy to search and work with data from the U.S.
