### 2022-03-14 #505 @maelle
:wave: @frankfarach Thanks for your submission! I'll start looking for reviewers, I have just one question and one comment below.

- [x **License:** The package has a CRAN or OSI accepted license.
---

- There aren't many functions but you could still add grouping to the reference index so that the most important functions come first.
- I was curious to know why you use `mockery::.mockPaths()`?

---
