### 2022-03-14 ropensci/npi#45 @frankfarach
Group the reference index to highlight the most important functions
Per https://github.com/ropensci/software-review/issues/505#issuecomment-1066447979
