+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   npi: npi.html
-last_built: 2022-03-04T06:14Z
+last_built: 2022-03-14T14:46Z
 
