### 2022-03-14 ropensci/npi#46 @frankfarach
Clean up references to mockery vs. httptest in testthat setup
1. Remove reference to httptest::.mockPaths() if not needed, or justify it ([line](https://github.com/frankfarach/npi/blob/0476eda909405af7100f18f7a57263f06dd94751/tests/testthat/setup.R#L4)). See  this [rOpenSci editorial comment](https://github.com/ropensci/software-review/issues/505#issuecomment-1066447979).

2. Remove reference to mockery package if not needed ([line](https://github.com/frankfarach/npi/blob/0476eda909405af7100f18f7a57263f06dd94751/tests/testthat/setup.R#L2)).
