### 2022-03-14 #505 @frankfarach
Hi @maelle, thanks for the comment and question, which I address below.

> #### Editor comments
> * There aren't many functions but you could still add grouping to the reference index so that the most important functions come first.

Thanks for the suggestion. I have created an [issue](https://github.com/frankfarach/npi/issues/45) for this.

> * I was curious to know why you use `mockery::.mockPaths()`?

[`.mockPaths()`](https://github.com/nealrichardson/httptest/blob/5048d8e88c3bcdafdfca56b52b4dd9d8578ea195/R/mock-paths.R) is from the `httptest` package. It lets you set a directory other than the current working directory for recording API fixtures with functions like `httptest::with_mock_api()`.

Your question made me realize that I am loading both `httptest` and `mockery` in tests/testthat/setup.R. I think I am only using functions from `httptests` in my test suite, so I may not need `mockery` at all. I've created another [issue](https://github.com/frankfarach/npi/issues/46) to check on this as part of my response to peer review.
