### 2022-03-15 #505 @maelle
:wave: @frankfarach! Thank you! I realized afterwards `.mockPaths()` was from httptest, as I was working on a package that uses httptest. :woman_facepalming: (in that package we use `with_mock_dir()`)
