### 2022-03-17 e656dc2
Feature/advanced uses vignette (#43)
* added intro to npi vignette

* auto updated DESCRIPTION file and git ignore

* updated based on review comments

* added advanced uses vignette

* adv use vignette updates

* added toc to both vignettes

* updated based on feedback

* added sentence on motivation for section

Co-authored-by: parmsam <parmsam>
