### 2022-04-04 #505 @Rekyt
---

Overall the package is very well designed. It has simple and meaningful functions and hides a quite important machinery to make the results of the NPI API easily accessible and exploitable in R. I would like to congratulate the author for this work.

Regarding code efficiency I didn't found any improvement that could be easily made as I have the impression the code is pretty optimized both from a reading and performance standpoint.

Regarding the documentation, I think it's very well written and clear. It's easy to understand it reading from different entry points. The difficulties I had understanding were rather related to the NPI API that I am not familiar with.

I have however a few comments on how to improve the coverage of the package and found several edge cases not well covered by the package, that could be worth to include as additional tests.

**API URL**

You could indicate API URL in DESCRIPTION file, this is needed for submission to CRAN (if you consider it). It is also possible to specify which link is what in the same file. Like it is done in `taxize` DESCRIPTION file https://github.com/ropensci/taxize/blob/a4db9a790e48c93235cac6145826c097ec687408/DESCRIPTION#L10-L12

**`npis` description**

The description of `npis` dataset doesn't seem to fit the dataset as it specifies "list of 0-n tibbles" for one column. It is not clear to me how to read it. Though I think it's nice to indicate the type of element. Maybe a way to make it more explicit would be to indicate "list-column of tibbles with X rows and Y columns".

**extra vignette: use cases**

I have no idea how this API could be useful in a real use case. But it may be nice to have an extra vignette showing how you can answer some questions through this API (how is it useful for research? for health administration?).

### Function usage

**Rate limitation**

Is there a form of rate limitation? It seems reasonable given that we access an API. Nothing is shown from the API standpoint nor the package? (I've seen it mentioned in the ["Getting started" vignette](https://frankfarach.github.io/npi/articles/npi.html#increasing-number-of-records-returned) but it could be more explicit). If it is the case, that could be a nice thing to mention to the user so that they are aware of the time it would take to get their results back.

**Can't use `npi_summarize()` with country code**

I've found an issue when querying by country code which returns slightly differently formatted query and as such renders `npi_summarize()` unusable.

reprex:

```r
library("npi")
ki = npi_search(country_code = "DE")
#> Requesting records 0-10...
npi_summarize(ki)
#> Error:
#> ! Tibble columns must have compatible sizes.
#> * Size 10: Existing data.
#> * Size 12: Column `primary_taxonomy`.
#> i Only values of size one are recycled.
#> Run `rlang::last_error()` to see where the error occurred.
```

Maybe `npi_summarize()` should take into account this edge case (even if I suppose there are not that many American health professionals outside the US).

**Error when internet is off**

Because this is a recurrent theme for API packages (and CRAN asks them to "fail gracefully" in these cases), I've checked what
happens when I turn of the internet and I don't think it is that explicit for the user.

```
> npi_search(city = "San Francisco")
Requesting records 0-10...
Error in curl::curl_fetch_memory(url, handle = handle) :
  Could not resolve host: npiregistry.cms.hhs.gov
```

Maybe there could be a way to display a better error message when internet is off or the website unreachable?
Like check first hand if internet is off and errors explicitly if it is the case.

**Special characters in `npi_search()` arguments**

Character arguments of `npi_search()` allow for some special character as specified by the documentation. But when searching with other special characters, the query is still submitted.

reprex:

``` r
library("npi")
npi_search(first_name = "KOŒ*")
#> Requesting records 0-10...
#> Error in `npi_handle_response()`:
#> !
#> Field: first_name
#> Field contains special character(s) or wrong number of characters
```

It works well but it fallback on the API when this is the case, I think it could be possible to catch these issues early when processing these arguments to save time and additional queries.

**Wildcards special cases**

Same as above, some special queries with wildcards used improperly (not trailing) could be caught earlier. I had strange results with some.

reprex:

``` r
library("npi")

# Trying to mess up with wildcards
ko = npi_search(last_name = "M*ll*")
#> Requesting records 0-10...
ko$basic  # The answer has nothing to do with the pattern
#> [[1]]
#> # A tibble: 1 x 11
#>   first_name last_name middle_name credential sole_proprietor gender
#>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
#> 1 JENNY      ENSTROM   E           PA         NO              F
#> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
#> #   status <chr>, name <chr>, certification_date <chr>

# Wildcards in the middle do not work
ko = npi_search(last_name = "M*ll")
#> Requesting records 0-10...
ko$basic  # And I get the same answer instead of an error?!
#> [[1]]
#> # A tibble: 1 x 11
#>   first_name last_name middle_name credential sole_proprietor gender
#>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
#> 1 JENNY      ENSTROM   E           PA         NO              F
#> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
#> #   status <chr>, name <chr>, certification_date <chr>
```

### User agent

It's great that the package provides a user-agent. It has been proven to be the best way possible to access an API to provide
a proof of who is accessing the API.

In the `README` there is a mention of the user agent (https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/README.Rmd#L114-L120). Which seems a great idea. However, this does not consider that the user may not be familiar with the concept of user-agent. Also, the displayed example doesn't show in which way it can be helpful to change its user-agent. Maybe you could go to a more concrete example. I also wonder if it is a good idea to let the user entirely remove the reference to the package.

To improve the user-agent, which only points to the [URL of the repo](https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/R/utils.R#L4), it would be good to indicate the version `npi` in use like it is done in [`taxize`](https://github.com/ropensci/taxize/blob/7eaf9a63473d5a277a72a02ea7899ee2d7da52ea/R/zzz.R#L249). If it is needed for the user to modify its user agent, then maybe providing a fixed part of the UA with the version of the package and a customizable second part would be even more complete.

**`checkmate`**

`checkmate` is in `Suggests` but is used across several functions and no tests are done if package isn't installed. Maybe a good idea to make an `Imports`?

**`tidyr` warning**

At several occasion we face this message:

```
The `.sep` argument of `unnest()` is deprecated as of tidyr 1.0.0.
Use `names_sep = '_'` instead.
```

while there is a function to check for installed tidyr version
`tidyr_new_interface()`.
On my computer I had the error when using `npi_flatten()` even though I had
tidyr v.1.2.0 installed

Functions `delay_by()` and `remove_null()` are defined but never used (and never tested). Was that planned for another update? Because it artificially reduces the coverage.

No tests have been done with empty results? Certain parameter combinations give back empty results and it could be easy to
cover L177 of npi_search.R https://app.codecov.io/gh/frankfarach/npi/blob/master/R/npi_search.R (for ex. `npi_search(last_name = "Grenie")`;))

`npi_flatten()` is not covered at all and could be covered by developing tests using the `npis` dataset.
