### 2022-04-04 ropensci/npi#47 @frankfarach
Indicate API URL in DESCRIPTION file
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> You could indicate API URL in DESCRIPTION file, this is needed for submission to CRAN (if you consider it). It is also possible to specify which link is what in the same file. Like it is done in taxize DESCRIPTION file https://github.com/ropensci/taxize/blob/a4db9a790e48c93235cac6145826c097ec687408/DESCRIPTION#L10-L12
