### 2022-04-04 ropensci/npi#48 @frankfarach
Clarify description of npis dataset
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> The description of npis dataset doesn't seem to fit the dataset as it specifies "list of 0-n tibbles" for one column. It is not clear to me how to read it. Though I think it's nice to indicate the type of element. Maybe a way to make it more explicit would be to indicate "list-column of tibbles with X rows and Y columns".
