### 2022-04-04 ropensci/npi#49 @frankfarach
Add illustrative use case to vignette
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> I have no idea how this API could be useful in a real use case. But it may be nice to have an extra vignette showing how you can answer some questions through this API (how is it useful for research? for health administration?).
