### 2022-04-04 ropensci/npi#50 @frankfarach
Clarify API rate limitation
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> Is there a form of rate limitation? It seems reasonable given that we access an API. Nothing is shown from the API standpoint nor the package? (I've seen it mentioned in the ["Getting started" vignette](https://frankfarach.github.io/npi/articles/npi.html#increasing-number-of-records-returned) but it could be more explicit). If it is the case, that could be a nice thing to mention to the user so that they are aware of the time it would take to get their results back.

Do this in at least two places: README.md and vignettes
