### 2022-04-04 ropensci/npi#51 @frankfarach
Can't use `npi_summarize()` with country code
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> I've found an issue when querying by country code which returns slightly differently formatted query and as such renders npi_summarize() unusable.
>
> reprex:

```
library("npi")
ki = npi_search(country_code = "DE")
#> Requesting records 0-10...
npi_summarize(ki)
#> Error:
#> ! Tibble columns must have compatible sizes.
#> * Size 10: Existing data.
#> * Size 12: Column `primary_taxonomy`.
#> i Only values of size one are recycled.
#> Run `rlang::last_error()` to see where the error occurred.
```
> Maybe npi_summarize() should take into account this edge case (even if I suppose there are not that many American health professionals outside the US).
