### 2022-04-04 ropensci/npi#52 @frankfarach
Improve error message when internet off or endpoint unreachable
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> Because this is a recurrent theme for API packages (and CRAN asks them to "fail gracefully" in these cases), I've checked what happens when I turn of the internet and I don't think it is that explicit for the user.

```
> npi_search(city = "San Francisco")
Requesting records 0-10...
Error in curl::curl_fetch_memory(url, handle = handle) :
  Could not resolve host: npiregistry.cms.hhs.gov
```

> Maybe there could be a way to display a better error message when internet is off or the website unreachable?
> Like check first hand if internet is off and errors explicitly if it is the case.
