### 2022-04-04 ropensci/npi#53 @frankfarach
Handle illegal special characters in query parameters
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> Character arguments of npi_search() allow for some special character as specified by the documentation. But when searching with other special characters, the query is still submitted.
>
> reprex:

```
library("npi")
npi_search(first_name = "KOŒ*")
#> Requesting records 0-10...
#> Error in `npi_handle_response()`:
#> !
#> Field: first_name
#> Field contains special character(s) or wrong number of characters
```

> Created on 2022-04-04 by the [reprex package](https://reprex.tidyverse.org/) (v2.0.1)
>
> It works well but it fallback on the API when this is the case, I think it could be possible to catch these issues early when processing these arguments to save time and additional queries.
