### 2022-04-04 ropensci/npi#54 @frankfarach
Handle wildcard special cases
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> some special queries with wildcards used improperly (not trailing) could be caught earlier. I had strange results with some.
>
> reprex:

```
library("npi")

# Trying to mess up with wildcards
ko = npi_search(last_name = "M*ll*")
#> Requesting records 0-10...
ko$basic  # The answer has nothing to do with the pattern
#> [[1]]
#> # A tibble: 1 x 11
#>   first_name last_name middle_name credential sole_proprietor gender
#>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
#> 1 JENNY      ENSTROM   E           PA         NO              F
#> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
#> #   status <chr>, name <chr>, certification_date <chr>

# Wildcards in the middle do not work
ko = npi_search(last_name = "M*ll")
#> Requesting records 0-10...
ko$basic  # And I get the same answer instead of an error?!
#> [[1]]
#> # A tibble: 1 x 11
#>   first_name last_name middle_name credential sole_proprietor gender
#>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
#> 1 JENNY      ENSTROM   E           PA         NO              F
#> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
#> #   status <chr>, name <chr>, certification_date <chr>
```

> Created on 2022-04-04 by the [reprex package](https://reprex.tidyverse.org/) (v2.0.1)
