### 2022-04-04 ropensci/npi#55 @frankfarach
User agent documentation and customization
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> It's great that the package provides a user-agent. It has been proven to be the best way possible to access an API to provide
a proof of who is accessing the API.
>
> In the README there is a mention of the user agent (https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/README.Rmd#L114-L120). Which seems a great idea. However, this does not consider that the user may not be familiar with the concept of user-agent. Also, the displayed example doesn't show in which way it can be helpful to change its user-agent. Maybe you could go to a more concrete example. I also wonder if it is a good idea to let the user entirely remove the reference to the package.
>
> To improve the user-agent, which only points to the [URL of the repo](https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/R/utils.R#L4), it would be good to indicate the version npi in use like it is done in [taxize](https://github.com/ropensci/taxize/blob/7eaf9a63473d5a277a72a02ea7899ee2d7da52ea/R/zzz.R#L249). If it is needed for the user to modify its user agent, then maybe providing a fixed part of the UA with the version of the package and a customizable second part would be even more complete.
