### 2022-04-04 ropensci/npi#56 @frankfarach
Fix tidyr warning
Note: This is a bug related to #24.

From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> At several occasion we face this message:

```
Warning message:
The `.sep` argument of `unnest()` is deprecated as of tidyr 1.0.0.
Use `names_sep = '_'` instead.
while there is a function to check for installed tidyr version
tidyr_new_interface().
```

> On my computer I had the error when using npi_flatten() even though I had
tidyr v.1.2.0 installed
