### 2022-04-04 ropensci/npi#57 @frankfarach
Review suggesting vs. importing checkmate package
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> checkmate is in Suggests but is used across several functions and no tests are done if package isn't installed. Maybe a good idea to make an Imports?
