### 2022-04-04 ropensci/npi#58 @frankfarach
Remove unused and untested functions
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> Functions delay_by() and remove_null() are defined but never used (and never tested). Was that planned for another update? Because it artificially reduces the coverage.
