### 2022-04-04 ropensci/npi#59 @frankfarach
Add test for empty results of `npi_search()`
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

> No tests have been done with empty results? Certain parameter combinations give back empty results and it could be easy to cover L177 of npi_search.R https://app.codecov.io/gh/frankfarach/npi/blob/master/R/npi_search.R (for ex. npi_search(last_name = "Grenie");))
