### 2022-04-04 ropensci/npi#60 @frankfarach
Add @Rekyt as reviewer in DESCRIPTION
From https://github.com/ropensci/software-review/issues/505#issuecomment-1087579526:

>  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
