+++ R/extractors.R
@@ -95,10 +95,10 @@ get_list_col <- function(df, list_col = NULL, key = "npi") {
 
   # Make unnest work with older and newer versions of tidyr
   if (tidyr_new_interface()) {
-    # Code for tidyr v1.0.0
+    # Old interface
     tidyr::unnest(df, !!rlang::sym(list_col), .sep = sep_val)
   } else {
-    # Code for v0.8.3
+    # New interface
     tidyr::unnest(df, !!rlang::sym(list_col), names_sep = sep_val)
   }
 }
