### 2022-04-07 #505 @maelle
@zabore thank you! actually it was a great oversight as it allowed us to add the possibility to log a review without a time entered yet. :wink:
