### 2022-04-07 #505 @frankfarach
> Thanks a ton @Rekyt & @zabore for your thorough reviews! @zabore, would you mind adding the time spent reviewing to your comment after "Estimated hours spent reviewing: "? Thank you!

Likewise, @Rekyt and @zabore, thanks so much for your reviews! I really appreciate the time you took to review the package and look forward to responding.

> @frankfarach I think the backlinks & milestone are great. We will still expect your response to be more or less self-explanatory so not just a list of links (I'm not assuming you got this wrong, just stating this just in case!).

@maelle Understood. I'll submit a self-contained prose response. The links will only be there for convenience. Thanks!
