### 2022-04-09 ropensci/npi#63 @frankfarach
Add @zabore as reviewer in DESCRIPTION
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

>  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
