+++ DESCRIPTION
@@ -6,7 +6,9 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-2145-0145")),
     person("Sam", "Parmar", , "parmartsam@gmail.com", role = "ctb"),
     person("Matthias", "Grenié", , "matthias.grenie@idiv.de", role = "rev", 
-           comment = c(ORCID = "0000-0002-4659-7522"))
+           comment = c(ORCID = "0000-0002-4659-7522")),
+    person("Emily", "Zabor", middle = "C.", "zabore2@ccf.org", role = "rev", 
+           comment = c(ORCID = "0000-0002-1402-4498"))
   )
 Description: Access the United States National Provider Identifier
     Registry API (if available) and provide informative error messages
