### 2022-04-09 ropensci/npi#64 @frankfarach
README: Clarify that enumeration_type is returned, not provider_type
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> Under "Search the registry" section the README mentions 4 regular vector columns that are returned by `npi_search()` but when I execute the code I get `enumeration_type` in place of the mentioned `provider_type`
