### 2022-04-09 ropensci/npi#65 @frankfarach
README: Indicate that code results may not be reproducible
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:;

> The example in the README is not reproducible, meaning I got different results when I ran the code than what was shown in the README output. Also, the ouput included in the vignette differ from those in the README. I assume this is due to a change in the NPI itself between when the README was created and when I ran the code? If possible, reproducible example in the README is ideal, though I understand that may be a challenge with a changing API.
