### 2022-04-09 ropensci/npi#66 @frankfarach
vignette: Organize search arguments by name, not description
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> In the "Search registry" section I would prefer to see the other search arguments listed by the argument name rather than a description (i.e. to see first_name rather than "provider's first name") as this
