### 2022-04-09 ropensci/npi#67 @frankfarach
vignette: Detail what `npi_summarize()` does
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> More details could be provided about what `npi_summarize()` does. Instead of saying it provides a more "human-readable" overview you could instead just state explicitly that it extract the first and last name, primary taxonomy, and primary address and phone number for each provider.
