### 2022-04-09 ropensci/npi#68 @frankfarach
vignette: Detail what `npi_flatten()` does
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> The README contains slightly more detailed information about what `npi_flatten()` does than the vignette. For users familiar with `purrr::flatten()` the action may be somewhat self explanatory. But for other users the term "flatten" likely has no meaning, so a more functional description could be useful.
