### 2022-04-09 ropensci/npi#69 @frankfarach
README and vignette: Provide use case for npi_is_valid()
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> It's not clear to me how or when `npi_is_valid()` would be used. Could provide some motivation or an actual use case in the README and/or vignette.
