### 2022-04-09 frankfarach/npi#48 @frankfarach
See also https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> In the documentation for the npis dataset you mention in several places "[list of 0-n tibbles]" but there is no definition of n. I assume this indicates the number of tibbles in the list can vary, but I don't think it's necessary to include this. "List of tibbles" could be sufficient.
