### 2022-04-09 ropensci/npi#70 @frankfarach
npi_search: Add links to data dictionary for results
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> Is there documentation anywhere for what information is contained in the columns of the tibble returned by `npi_search()`? I couldn't find it.
