### 2022-04-09 ropensci/npi#71 @frankfarach
npi_search: Clarify how default matching works for arguments
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> The help page (or vignette?) for `npi_search()` could provide more details for how matching is done by default. For example, I get different results if I search for city = "new york" versus city = "new york city". I know you have not programmed the name matching yourself, but I think this is important information for the user to have so at least pointing to a link where this information can be found could be useful.
