### 2022-04-09 ropensci/npi#72 @frankfarach
npi_search: Clarify user message to indicate that a subset of records is being returned
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> It could be useful to add a message from npi_search() that explicitly states something along the lines of "Returning 10 of xxx records for the specified search parameters" so that the user is aware they are getting only a subset of the records that match their search parameters.
