### 2022-04-09 frankfarach/npi#53 @frankfarach
See also https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> It seems like when invalid arguments are passed to npi_search() it still tries to query the API first and then returns an error. Internal checks could be done in this function to avoid the query altogether in these cases and to more quickly return a useful error message.
