### 2022-04-09 ropensci/npi#73 @frankfarach
npi_flatten: Investigate and test edge cases
See https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> Does npi_flatten() always return a sensible result? It seems like it merges together a lot of different pieces of information from different list columns but does it always make sense to cross-tabulate every single one? I don't have an idea of where it might go wrong, but always wonder when merging tibbles with different numbers of rows.
