### 2022-04-09 frankfarach/npi#49 @frankfarach
See also https://github.com/ropensci/software-review/issues/505#issuecomment-1090531882:

> Overall I found this package to be well designed and well documented. Nice work! I understand what it does, but I was left wondering why. It could be nice to include a vignette demonstrating how you envision this being used. How are people getting this information now, and how will this streamline things for them?
