### 2022-04-10 4fdccf5
Closes #50 and #64
