+++ man/npi-package.Rd
@@ -40,9 +40,9 @@ by the `npi` column into an analysis-ready object.
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/frankfarach/npi}
-  \item \url{https://frankfarach.github.io/npi/}
-  \item \url{https://npiregistry.cms.hhs.gov/api/}
+  \item \url{https://github.com/frankfarach/npi (devel)}
+  \item \url{https://frankfarach.github.io/npi/ (user manual)}
+  \item \url{https://npiregistry.cms.hhs.gov/api/ (API)}
   \item Report bugs at \url{https://github.com/frankfarach/npi/issues}
 }
 
