### 2022-04-16 frankfarach/npi#52 @frankfarach
I was able to reproduce the reported issue and fix it. Here is a reprex after fixing it:

```
npi::npi_search(city = "San Francisco")
#> 10 records requested
#> Requesting records 0-10...
#> Error in `npi_api()` at npi/R/api.R:49:2:
#> ! Unable to access Internet. Check network connection and try again.
```
