### 2022-04-16 frankfarach/npi#59 @frankfarach
Added mock and test which now passes:
```
> library(httptest)
> start_capturing()
> npi_search(city = "ZZZZZZZ")
10 records requested
Requesting records 0-10...
# A tibble: 0 × 0
> stop_capturing()
```
