### 2022-04-16 frankfarach/npi#53 @frankfarach
The rules implemented by the API are described at https://npiregistry.cms.hhs.gov/registry/help-api. The unique sets of special characters to test are:

Set 1 (Used in: `first_name`, `last_name`, `city`)
ampersand, apostrophe, colon, comma, forward slash, hyphen, left and right parentheses, period, pound sign, quotation mark, and semi-colon

Set 2 (Used in: `organization_name`)
Set 1 plus "at" sign

Approach:

1. Create a validator function that takes a list of legal characters and an input string and returns TRUE if all of the input string's characters are in the set of legal characters, otherwise FALSE.
2. In `npi_search()`, for each non-empty argument that needs to be validated, call the validator function on it, passing in the legal set of characters for the field (see above).
3. If the validation check fails, return a friendly error and abort.
4. Add tests for each validation set.
