+++ R/npi_search.R
@@ -120,6 +120,16 @@ npi_search <- function(number = NULL,
 
   enumeration_type <- ifelse(enumeration_type == "ind", "NPI-1", "NPI-2")
 
+  # Check for illegal characters
+  legal_1 <- "[^[:alnum:][:space:]()&:,-.#;'/\"]"
+  legal_2 <- "[^[:alnum:][:space:]()&:,-.#;'/\"@]"
+
+  if (any(stringr::str_detect(c(first_name, last_name, city), legal_1)) ||
+      any(stringr::str_detect(organization_name, legal_2))) {
+    msg <- "Field contains at least one illegal character. See `?npi_search`."
+    rlang::abort(msg, class = "illegal_character")
+  }
+
   if (!is.logical(use_first_name_alias) &&
     !is.null(use_first_name_alias)) {
     rlang::abort("`use_first_name_alias` must be TRUE or FALSE if specified.")
