### 2022-04-17 frankfarach/npi#54 @frankfarach
Per https://npiregistry.cms.hhs.gov/registry/help-api, "Trailing wildcard entries are permitted requiring at least two characters to be entered (e.g. "jo*" )" applies to `first_name`, `last_name`, `organization_name`, and `postal_code`.
