+++ R/utils.R
@@ -1,7 +1,8 @@
 # Global variables for GET request
 API_VERSION <- "2.1" # Referenced in `npi_search()`
 BASE_URL <- "https://npiregistry.cms.hhs.gov/api/"
-USER_AGENT <- "http://github.com/frankfarach/npi"
+USER_AGENT <- paste(paste0("npi/", utils::packageVersion("npi")),
+                    "(http://github.com/frankfarach/npi)")
 MAX_N_PER_REQUEST <- 200L
 
 #' Handle bad function arguments
