+++ tests/testthat/test-api.R
@@ -3,7 +3,8 @@ context("test-api.R")
 # Global variables for GET request
 API_VERSION <- "2.1"
 BASE_URL <- "https://npiregistry.cms.hhs.gov/api/"
-USER_AGENT <- "http://github.com/frankfarach/npi"
+USER_AGENT <- paste(paste0("npi/", utils::packageVersion("npi")),
+                    "(http://github.com/frankfarach/npi)")
 MAX_N_PER_REQUEST <- 200L
 
 
