### 2022-04-24 #505 @frankfarach
Hi @maelle, I wanted to let you know that I'm still making changes and preparing my response. You can see my commits [here](https://github.com/frankfarach/npi/tree/feature/ropensci-peer-review-1). I've addressed about 2/3 of the issues raised by reviewers so far. Based on my slow-but-steady progress to date and my available time in near future, I am aiming to submit my response by May 6. I mention this as I know there's an expectation for package authors to provide their responses within two weeks of the last-submitted review, and we're a few days beyond that period already. Please let me know if my proposed timeline works for you. Thanks for your patience!
