### 2022-04-26 #505 @maelle
@frankfarach you should now be able to add a comment with the contents `@ropensci-review-bot check package` in order to get a pkgcheck report :-) (it was previously only possible for editors)
