### 2022-08-22 frankfarach/npi#74 @trang-n
Error with npi_search
Hi @frankfarach, thanks for a very useful package!
I encountered this error today while running npi_search. Would appreciate any insights!

```
> nyc <- npi_search(city = "New York City")
Requesting records 0-10...
Error in `npi_handle_response()`:
! API returned text/plain, not JSON.
Run `rlang::last_error()` to see where the error occurred.
```
