### 2022-08-23 frankfarach/npi#74 @trang-n
I investigated this more today and looks like they changed the API output format to text.
