### 2022-08-24 frankfarach/npi#74 @frankfarach
Hi @trang-n, thanks for identifying this issue! I was able to replicate the problem and agree with your diagnosis. I will dig into this further in the next few days and hopefully get the issue resolved soon.
