### 2022-09-08 #505 @frankfarach
> 👋 @frankfarach, any update? 😸

The API recently introduced a [change in the response type](https://github.com/frankfarach/npi/issues/74) that I need to fix as it prevents all search requests from being parsed. I am hoping to have this resolved in the next week, then I will return to addressing the remaining issues mentioned by the reviewers.
