### 2022-09-28 frankfarach/npi#74 @parmsam
@frankfarach, added possible solution in PR.
