### 2022-09-28 ed06194
handle case of 10-digit epoch format
