### 2022-10-22 99b04c3
Update epoch field conversion to date (#75)
* update epoch field conversion to date

* handle case of 10-digit epoch format
