+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.2
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   npi: npi.html
-last_built: 2022-03-14T14:46Z
+last_built: 2022-10-22T00:28Z
 
