### 2022-10-22 #505 @frankfarach
@maelle The API issue has been fixed and was pushed to master as a hotfix today, so the package is useable again. I am incorporating the hotfix into the branch where I'm working on package updates to be summarized in my forthcoming responses to the reviewers' comments. I'll send an update in a few days once I get my bearings again on the work that's left to do. Thank you very much for your patience with the delay!
