### 2022-10-23 4ac53d1
Motivate the use of npi_is_valid() (Closes #69)
