+++ README.md
@@ -62,6 +62,7 @@ nyc <- npi_search(city = "New York City")
 ```
 
 ``` r
+# Your results may differ since the data in the NPPES database changes over time
 nyc
 #> # A tibble: 10 × 11
 #>       npi enumeration_type basic    other_names identifiers taxonomies addresses
