### 2022-10-23 frankfarach/npi#46 @frankfarach
mockery::stub is used to stub npi_handle_response() to test the handling of a bad response in one of the [tests](https://github.com/frankfarach/npi/blob/99b04c3936ef2f46276b02046b137cd5a179b1d0/tests/testthat/test-api.R#L52).
