### 2022-10-23 frankfarach/npi#71 @frankfarach
This is a great question, but I am unable to find an answer in the API documentation online. If I can find an answer I will provide a link here.
