+++ README.Rmd
@@ -79,7 +79,7 @@ If you're comfortable [working with list columns](https://r4ds.had.co.nz/many-mo
 
 ## Working with search results 
 
-Run `npi_summarize()` on your results to see a more human-readable overview of what we've got:
+Run `npi_summarize()` on your results to see a more human-readable overview of what we've got. Specifically, the function returns the NPI number, provider's name, enumeration type (individual or organizational provider), primary address, phone number, and primary taxonomy (area of practice):
 
 ```{r summarize-nyc}
 npi_summarize(nyc)
