+++ README.Rmd
@@ -75,33 +75,40 @@ The full search results have four regular vector columns, `npi`, `enumeration_ty
 
 A full list of the possible fields within these list columns can be found on the [NPPES API Help page](https://npiregistry.cms.hhs.gov/registry/Json-Conversion-Field-Map).
 
-If you're comfortable [working with list columns](https://r4ds.had.co.nz/many-models.html), this may be all you need from the package. But let's not stop just yet, because `npi` provides convenience functions to summarize and extract the data you need.
+If you're comfortable [working with list columns](https://r4ds.had.co.nz/many-models.html), this may be all you need from the package. However, `npi` also provides functions that can help you summarize and transform your search results.
 
-## Working with search results 
+## Working with search results
 
-Run `npi_summarize()` on your results to see a more human-readable overview of what we've got. Specifically, the function returns the NPI number, provider's name, enumeration type (individual or organizational provider), primary address, phone number, and primary taxonomy (area of practice):
+`npi` has two main helper functions for working with search results: `npi_summarize()` and `npi_flatten()`.
+
+### Summarizing results
+
+Run `npi_summarize()` on your results to see a more human-readable overview of your search results. Specifically, the function returns the NPI number, provider's name, enumeration type (individual or organizational provider), primary address, phone number, and primary taxonomy (area of practice):
 
 ```{r summarize-nyc}
 npi_summarize(nyc)
 ```
 
-Suppose we just want the basic and taxonomy information for each NPI in the result in a flattened data frame:
+### Flattening results 
 
-```{r flatten-two}
-npi_flatten(nyc, c("basic", "taxonomies"))
-```
+As seen above, the data frame returned by `npi_search()` has a nested structure. Although all the data in a single row relates to one NPI, each list column contains a list of one or more values corresponding to the NPI for that row. For example, a provider's NPI record may have multiple associated addresses, phone numbers, taxonomies, and other attributes, all of which is stored in a single row of the data frame.
 
-Or we can flatten the whole thing and prune back later:
+Because nested structures can be a little tricky to work with, the `npi` includes `npi_flatten()`, a function that transforms the data frame into a flatter (i.e., unnested) structure that's easier to use. `npi_flatten()` performs the following transformations:
+
+  * unnest the list columns
+  * prefix the name of each unnested column with the name of its original list column
+  * join the data together by NPI
 
 ```{r flatten-all}
 npi_flatten(nyc)
 ```
 
-Now we're ready to do whatever else we need to do with this data. Under the hood, `npi_flatten()` has done a lot of data wrangling for us:
+If you only want to flatten a subset of the original data frame, you can optionally pass in a vector containing the list columns you want to unnest:
 
-  * unnested the specified list columns
-  * avoided potential naming collisions by prefixing the unnested names by their originating column name
-  * joined the data together by NPI
+```{r flatten-two}
+npi_flatten(nyc, c("basic", "taxonomies"))
+```
+
 
 ### Validating NPIs
 
@@ -114,6 +121,7 @@ npi_is_valid(1234567898)
 ```
 
 Note that this function doesn't check whether the NPI numbers are activated or deactivated (see [#22](https://github.com/frankfarach/npi/issues/22#issuecomment-787642817)). It merely checks for the number's consistency with the NPI specification. As such, it can help you detect and handle data quality issues early.
+
 
 ## Set your own user agent
 
