### 2022-10-31 #505 @frankfarach
@maelle I've addressed almost all of the comments in my commits to the package and have just a few tests left to write. I have started composing my written responses, which will elaborate on and reference the changes in a self-contained way. I should be finished in a couple of weeks.
