### 2022-11-02 3653899
Cover npi_flatten() with unit tests for #73
