### 2022-11-02 0565f4d
Closes #73. Add guidance on npi_flatten() to README.
