### 2022-11-03 #505 @frankfarach
### Response to @Rekyt

> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * **Briefly describe any working relationship you have (had) with the package authors.**
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s):** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions
> * [x]  **Examples:** (that run successfully locally) for all exported functions
> * [x]  **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
>
> Estimated hours spent reviewing: 3
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Your contribution as a package reviewer has been [acknowledged](https://github.com/frankfarach/npi/commit/33c458d2b46ef7e9663dbed93e5d3db2251f3b1c) in the DESCRIPTION file. Please verify that your ORCID is correct and/or whether you want it referenced.

> ### Review Comments
> Overall the package is very well designed. It has simple and meaningful functions and hides a quite important machinery to make the results of the NPI API easily accessible and exploitable in R. I would like to congratulate the author for this work.
>
> Regarding code efficiency I didn't found any improvement that could be easily made as I have the impression the code is pretty optimized both from a reading and performance standpoint.
>
> Regarding the documentation, I think it's very well written and clear. It's easy to understand it reading from different entry points. The difficulties I had understanding were rather related to the NPI API that I am not familiar with.
>
> I have however a few comments on how to improve the coverage of the package and found several edge cases not well covered by the package, that could be worth to include as additional tests.

Thank you for the detailed and constructive feedback, @Rekyt, and thank you for your patience! Please see my responses in-line below, and let me know if any further clarification or update is needed.

> ### Documentation
> **API URL**
>
> You could indicate API URL in DESCRIPTION file, this is needed for submission to CRAN (if you consider it). It is also possible to specify which link is what in the same file. Like it is done in `taxize` DESCRIPTION file https://github.com/ropensci/taxize/blob/a4db9a790e48c93235cac6145826c097ec687408/DESCRIPTION#L10-L12

[Done](https://github.com/frankfarach/npi/commit/db307c7cdd735922585c941617f1d202d4d10abb):

```
URL: https://github.com/frankfarach/npi (devel),
https://frankfarach.github.io/npi/ (user manual),
https://npiregistry.cms.hhs.gov/api/ (API)
```

> **`npis` description**
>
> The description of `npis` dataset doesn't seem to fit the dataset as it specifies "list of 0-n tibbles" for one column. It is not clear to me how to read it. Though I think it's nice to indicate the type of element. Maybe a way to make it more explicit would be to indicate "list-column of tibbles with X rows and Y columns".

The documentation now [refers](https://github.com/frankfarach/npi/commit/0ab6e7c2f19d530ab1a83bfad2a1201b19c7ae3b) to "list of tibbles" instead of "list of 0-n tibbles".

> **extra vignette: use cases**
>
> I have no idea how this API could be useful in a real use case. But it may be nice to have an extra vignette showing how you can answer some questions through this API (how is it useful for research? for health administration?).

The README now [provides further context](https://github.com/frankfarach/npi/commit/640b145e6a733e78034158ff359247f766c95d6a) on motivation and usage:
```
Analysts working with healthcare and public health data frequently need to join data from multiple sources to answer their business or research questions. Unfortunately, joining data in healthcare is hard because so few entities have unique, consistent identifiers across organizational boundaries. NPI numbers, however, do not suffer from these limitations, as all U.S. providers meeting certain common criteria must have an NPI number in order to be reimbursed for the services they provide. This makes NPI numbers incredibly useful for joining multiple datasets by provider, which is the primary motivation for developing this package.
```

> ### Function usage
> **Rate limitation**
>
> Is there a form of rate limitation? It seems reasonable given that we access an API. Nothing is shown from the API standpoint nor the package? (I've seen it mentioned in the ["Getting started" vignette](https://frankfarach.github.io/npi/articles/npi.html#increasing-number-of-records-returned) but it could be more explicit). If it is the case, that could be a nice thing to mention to the user so that they are aware of the time it would take to get their results back.

The `npi` vignette now [states](https://github.com/frankfarach/npi/commit/4fdccf56b373bf07d7c8f684dc73abf221cbe76d):
```
The NPPES API documentation does not specify additional API rate limitations. However, if you need more than 1200 NPI records for a set of search terms, you will need to download the [NPPES Data Dissemination File](https://download.cms.gov/nppes/NPI_Files.html).
```

> **Can't use `npi_summarize()` with country code**
>
> I've found an issue when querying by country code which returns slightly differently formatted query and as such renders `npi_summarize()` unusable.
>
> reprex:
>
> ```r
> library("npi")
> ki = npi_search(country_code = "DE")
> #> Requesting records 0-10...
> npi_summarize(ki)
> #> Error:
> #> ! Tibble columns must have compatible sizes.
> #> * Size 10: Existing data.
> #> * Size 12: Column `primary_taxonomy`.
> #> i Only values of size one are recycled.
> #> Run `rlang::last_error()` to see where the error occurred.
> ```
>
> Maybe `npi_summarize()` should take into account this edge case (even if I suppose there are not that many American health professionals outside the US).

Thanks for catching this and providing helpful details! Upon further investigation, some NPI records have multiple tibbles in the taxonomies list column of the results with `FALSE` for all values of the `primary` column. In that situation, there is no record with a primary taxonomy. The [fix](https://github.com/frankfarach/npi/commit/5a539a0f686b47624fdb4c44ae0c75c767cc5acc) was to return the first row of the taxonomies tibble grouped by NPI. This removes the error by ensuring a 1-to-1 relationship between NPIs and taxonomy rows.

``` r
library("npi")
ki = npi_search(country_code = "DE")
#> 10 records requested
#> Requesting records 0-10...
npi_summarize(ki)
#> # A tibble: 10 × 6
#>    npi        name      enumeration_type primary_practic… phone primary_taxonomy
#>    <chr>      <chr>     <chr>            <chr>            <chr> <chr>
#>  1 1316925936 PHYLLIS … Individual       <NA>             0972… Registered Nurse
#>  2 1306903968 SHANNON … Individual       <NA>             4962… Social Worker, …
#>  3 1124185780 SHAWN WH… Individual       <NA>             4962… Physician Assis…
#>  4 1679551931 JOSEPH P… Individual       <NA>             253-… Registered Nurs…
#>  5 1083761985 JAY HARDY Individual       CMR 414 BOX 139… 0114… Registered Nurs…
#>  6 1043367949 JACQUELI… Individual       HEIDELBERG MEDD… 4906… Licensed Practi…
#>  7 1316094493 QEVSERE … Individual       <NA>             0615… Social Worker, …
#>  8 1811044928 JULIE SO… Individual       <NA>             637-… Social Worker, …
#>  9 1083761035 PATRICIA… Individual       <NA>             0221… Registered Nurse
#> 10 1942288964 JEFFERY … Individual       USAMEDDAC WUERZ… 0114… Nurse Anestheti…
```

> **Error when internet is off**
>
> Because this is a recurrent theme for API packages (and CRAN asks them to "fail gracefully" in these cases), I've checked what happens when I turn of the internet and I don't think it is that explicit for the user.
>
> ```
> > npi_search(city = "San Francisco")
> Requesting records 0-10...
> Error in curl::curl_fetch_memory(url, handle = handle) :
>   Could not resolve host: npiregistry.cms.hhs.gov
> ```
>
> Maybe there could be a way to display a better error message when internet is off or the website unreachable? Like check first hand if internet is off and errors explicitly if it is the case.

I have added a [check](https://github.com/frankfarach/npi/commit/787f794e0b8dcb123815a9bec405141ba10c062c) for internet connectivity inside the internal `npi_api()` function using `curl::has_internet()` and aborting with a message upon failure:
```
  if (!curl::has_internet()) {
    msg <- "Unable to access Internet. Check network connection and try again."
    rlang::abort("internet_error", message = msg)
  }
```
``` r
library("npi")
npi_search(city = "San Francisco")
#> 10 records requested
#> Requesting records 0-10...
#> Error in `npi_api()` at npi/R/api.R:49:2:
#> ! Unable to access Internet. Check network connection and try again.
```

> **Special characters in `npi_search()` arguments**
>
> Character arguments of `npi_search()` allow for some special character as specified by the documentation. But when searching with other special characters, the query is still submitted.
>
> reprex:
>
> ```r
> library("npi")
> npi_search(first_name = "KOŒ*")
> #> Requesting records 0-10...
> #> Error in `npi_handle_response()`:
> #> !
> #> Field: first_name
> #> Field contains special character(s) or wrong number of characters
> ```
>
> Created on 2022-04-04 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)
>
> It works well but it fallback on the API when this is the case, I think it could be possible to catch these issues early when processing these arguments to save time and additional queries.

This issue has been [fixed](https://github.com/frankfarach/npi/commit/344ae6177df6057909fe9bd8abaf95de28fe9266) along the lines you suggested (i.e., catching "illegal" characters on the client side before making an API call).

> **Wildcards special cases**
>
> Same as above, some special queries with wildcards used improperly (not trailing) could be caught earlier. I had strange results with some.
>
> reprex:
>
> ```r
> library("npi")
>
> # Trying to mess up with wildcards
> ko = npi_search(last_name = "M*ll*")
> #> Requesting records 0-10...
> ko$basic  # The answer has nothing to do with the pattern
> #> [[1]]
> #> # A tibble: 1 x 11
> #>   first_name last_name middle_name credential sole_proprietor gender
> #>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
> #> 1 JENNY      ENSTROM   E           PA         NO              F
> #> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
> #> #   status <chr>, name <chr>, certification_date <chr>
>
> # Wildcards in the middle do not work
> ko = npi_search(last_name = "M*ll")
> #> Requesting records 0-10...
> ko$basic  # And I get the same answer instead of an error?!
> #> [[1]]
> #> # A tibble: 1 x 11
> #>   first_name last_name middle_name credential sole_proprietor gender
> #>   <chr>      <chr>     <chr>       <chr>      <chr>           <chr>
> #> 1 JENNY      ENSTROM   E           PA         NO              F
> #> # ... with 5 more variables: enumeration_date <chr>, last_updated <chr>,
> #> #   status <chr>, name <chr>, certification_date <chr>
> ```
>
> Created on 2022-04-04 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)
>

The package implements error-checking logic that is described by the API documentation with sufficient clarity. For example, since receiving reviewers' comments, the package now [checks for illegal characters](https://github.com/frankfarach/npi/commit/344ae6177df6057909fe9bd8abaf95de28fe9266) as those rules are clearly spelled out in the API documentation. The checks are performed for applicable user-supplied query parameters (i.e., `first_name`, `last_name`, `city`, `organization_name`) before sending the request to the API. Several unit tests have been added as well.

> ### User agent
> It's great that the package provides a user-agent. It has been proven to be the best way possible to access an API to provide a proof of who is accessing the API.
>
> In the `README` there is a mention of the user agent (https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/README.Rmd#L114-L120). Which seems a great idea. However, this does not consider that the user may not be familiar with the concept of user-agent. Also, the displayed example doesn't show in which way it can be helpful to change its user-agent. Maybe you could go to a more concrete example. I also wonder if it is a good idea to let the user entirely remove the reference to the package.
>
> To improve the user-agent, which only points to the [URL of the repo](https://github.com/frankfarach/npi/blob/d4e98a52ccc0a7f71328582333e5b4191f4796b9/R/utils.R#L4), it would be good to indicate the version `npi` in use like it is done in [`taxize`](https://github.com/ropensci/taxize/blob/7eaf9a63473d5a277a72a02ea7899ee2d7da52ea/R/zzz.R#L249). If it is needed for the user to modify its user agent, then maybe providing a fixed part of the UA with the version of the package and a customizable second part would be even more complete.

Thanks for the excellent suggestion! The default user agent now [references](https://github.com/frankfarach/npi/commit/0a9f9549e06bf315423307d87bbf4c7ab66bc5ea) the package name, current package version (as installed in the user's R session), and the URL to the GitHub repo for the package's source code. The README also explains this (see the same commit at the link above) and demonstrates how users can set their own optional user agent:

```
A [user agent](https://en.wikipedia.org/wiki/User_agent) is a way for the software interacting with an API to tell it who or what is making the request. This helps the API's maintainers understand what systems are using the API. By default, when `npi` makes a request to the NPPES API, the request header references the name of the package and the URL for the repository (e.g., '`r paste(paste0("npi/", utils::packageVersion("npi")), "(http://github.com/frankfarach/npi)")`'). If you want to set a custom user agent, update the value of the `npi_user_agent` option. For example, for version 1.0.0 of an app called "my_app", you could run the following code:
```
```
{r, eval = FALSE}
options(npi_user_agent = "my_app/1.0.0")
 ```
> ### Code
> **`checkmate`**
>
> `checkmate` is in `Suggests` but is used across several functions and no tests are done if package isn't installed. Maybe a good idea to make an `Imports`?

[Done](https://github.com/frankfarach/npi/commit/c89eae08a587364ad5f58cf793e973472a6a76b7)

> **`tidyr` warning**
>
> At several occasion we face this message:
>
> ```
> Warning message:
> The `.sep` argument of `unnest()` is deprecated as of tidyr 1.0.0.
> Use `names_sep = '_'` instead.
> ```
>
> while there is a function to check for installed tidyr version `tidyr_new_interface()`. On my computer I had the error when using `npi_flatten()` even though I had tidyr v.1.2.0 installed

Thanks to your comment and the `tidyr` [documentation](https://tidyr.tidyverse.org/articles/in-packages.html#conditional-code), the package now [gracefully handles](https://github.com/frankfarach/npi/commit/8f65260fadc9cd4998f915cf4fe991580bac8769) the interface change and should work regardless of the version of `tidyr` the user has installed.

> ### Tests
> Functions `delay_by()` and `remove_null()` are defined but never used (and never tested). Was that planned for another update? Because it artificially reduces the coverage.

You are correct. The functions have been [removed](https://github.com/frankfarach/npi/commit/c15c3deed84d9407ec283f57e2d2e7e9542c27f8), which indeed improved the test coverage. :)

> No tests have been done with empty results? Certain parameter combinations give back empty results and it could be easy to cover L177 of npi_search.R https://app.codecov.io/gh/frankfarach/npi/blob/master/R/npi_search.R (for ex. `npi_search(last_name = "Grenie")`;))

I've [added](https://github.com/frankfarach/npi/commit/4c47e66d434055140c6368a28b1efee3db26785c) a mock for 0 results and a corresponding test. Thanks for the suggestion!

> `npi_flatten()` is not covered at all and could be covered by developing tests using the `npis` dataset.

`npi_flatten()` is now [covered](https://github.com/frankfarach/npi/commit/365389993e8b84d6c3154dfe801064854187a235) with tests using the `npis` dataset.
