### 2022-11-03 #505 @frankfarach
### Response to @zabore

> Here is my review!
>
> * **Briefly describe any working relationship you have (had) with the package authors.**
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s):** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions
> * [x]  **Examples:** (that run successfully locally) for all exported functions
> * [x]  **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
>
> Estimated hours spent reviewing: 4
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Your contribution as a reviewer has been [acknowledged](https://github.com/frankfarach/npi/commit/0f946a60f6436128f5c1633a7825f00fdcf38f1c) accordingly in DESCRIPTION.

> ### Review Comments
> Overall I found this package to be well designed and well documented. Nice work! I understand what it does, but I was left wondering why. It could be nice to include a vignette demonstrating how you envision this being used. How are people getting this information now, and how will this streamline things for them? Below are some more specific comments, but overall I found no major issues.

Thank you, @zabore, for your thoughtful and detailed review. Please see my responses inline below, and let me know if I can clarify or update anything.

> #### README
> * Overall the README is thorough and well organized and easy to read. Great work!

Thanks!

> * Under "Search the registry" section the README mentions 4 regular vector columns that are returned by `npi_search()` but when I execute the code I get `enumeration_type` in place of the mentioned `provider_type`

Thanks for catching this. The README has been updated to reflect this change (see line 63 of the README.Rmd in [this commit](https://github.com/frankfarach/npi/commit/4fdccf56b373bf07d7c8f684dc73abf221cbe76d#))

> * When I ran `npi_summarize(nyc)` I got the following warning message (Note I have tidyr 1.2.0 installed on my machine):
>   ```
>   Warning message:
>   The `.sep` argument of `unnest()` is deprecated as of tidyr 1.0.0.
>   Use `names_sep = '_'` instead.
>   ```

Thanks to your comment and the `tidyr` [documentation](https://tidyr.tidyverse.org/articles/in-packages.html#conditional-code), the package now [gracefully handles](https://github.com/frankfarach/npi/commit/8f65260fadc9cd4998f915cf4fe991580bac8769) the interface change and should work regardless of the version of `tidyr` the user has installed.

> * The example in the README is not reproducible, meaning I got different results when I ran the code than what was shown in the README output. Also, the ouput included in the vignette differ from those in the README. I assume this is due to a change in the NPI itself between when the README was created and when I ran the code? If possible,  reproducible example in the README is ideal, though I understand that may be a challenge with a changing API.

The results returned by `npi_search()` in the README are those returned by the API itself at the time the document is compiled. The README now [clarifies](https://github.com/frankfarach/npi/commit/709feddf92741a62e73a394f3a440c5a0b7ab142) that "Your results may differ since the data in the NPPES database changes over time."

> #### vignette
> * In the "Search registry" section I would prefer to see the other search arguments listed by the argument name rather than a description (i.e. to see `first_name` rather than "provider's first name") as this

The vignette has been [updated](https://github.com/frankfarach/npi/commit/a3f338f59cb3f3bdb1fa0941b7e546fee021498c) to reference the individual argument names instead of their descriptions.

> * More details could be provided about what `npi_summarize()` does. Instead of saying it provides a more "human-readable" overview you could instead just state explicitly that it extract the first and last name, primary taxonomy, and primary address and phone number for each provider.

The README now [states](https://github.com/frankfarach/npi/commit/2bb17158b7a4c42751c04ba6dfa2f6a2f4e07441): "Specifically, the function returns the NPI number, provider's name, enumeration type (individual or organizational provider), primary address, phone number, and primary taxonomy (area of practice)."

> * The README contains slightly more detailed information about what `npi_flatten()` does than the vignette. For users familiar with `purrr::flatten()` the action may be somewhat self explanatory. But for other users the term "flatten" likely has no meaning, so a more functional description could be useful.

Additional paragraphs have been [added](https://github.com/frankfarach/npi/commit/45ed43aa8ec44d524604e3781f6843758d4c2a01) to the README to describe what `npi_flatten()` does:

"As seen above, the data frame returned by `npi_search()` has a nested structure. Although all the data in a single row relates to one NPI, each list column contains a list of one or more values corresponding to the NPI for that row. For example, a provider's NPI record may have multiple associated addresses, phone numbers, taxonomies, and other attributes, all of which is stored in a single row of the data frame.

Because nested structures can be a little tricky to work with, the `npi` includes `npi_flatten()`, a function that transforms the data frame into a flatter (i.e., unnested) structure that's easier to use. `npi_flatten()` performs the following transformations:

  * unnest the list columns
  * prefix the name of each unnested column with the name of its original list column
  * join the data together by NPI"

> #### Documentation/Examples
> * Great use of `npi_` to start all exported functions, makes them easy to use and get through autofill.

Thanks! I appreciate this when other packages do it. It helps with package maintenance, too.

> * It's not clear to me how or when `npi_is_valid()` would be used. Could provide some motivation or an actual use case in the README and/or vignette.

The README now [explains](https://github.com/frankfarach/npi/commit/4ac53d190f3ce470339f27c84ad5c212b596b1a4) how to use `npi_is_valid()` for error-checking: "Just like credit card numbers, NPI numbers can be mistyped or corrupted in transit. Likewise, officially-issued NPI numbers have a [check digit](https://en.wikipedia.org/wiki/Check_digit) for error-checking purposes. Use `npi_is_valid()` to check whether an NPI number you've encountered is [validly constructed](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPIcheckdigit.pdf):

```{r valid_npi_ex}
# Validate off NPIs
npi_is_valid(1234567893)
npi_is_valid(1234567898)
```

Note that this function doesn't check whether the NPI numbers are activated or deactivated (see [#22](https://github.com/frankfarach/npi/issues/22#issuecomment-787642817)). It merely checks for the number's consistency with the NPI specification. As such, it can help you detect data quality issues early."

> * In the documentation for the `npis` dataset you mention in several places "[list of 0-n tibbles]" but there is no definition of n. I assume this indicates the number of tibbles in the list can vary, but I don't think it's necessary to include this. "List of tibbles" could be sufficient.

The documentation now [refers](https://github.com/frankfarach/npi/commit/0ab6e7c2f19d530ab1a83bfad2a1201b19c7ae3b) to "list of tibbles" instead of "list of 0-n tibbles".

> * Is there documentation anywhere for what information is contained in the columns of the tibble returned by `npi_search()`? I couldn't find it.

The NPPES API documentation contains a data dictionary, which is now [referenced](https://github.com/frankfarach/npi/commit/389c5e623d7d081f63d40fc327ce17aa73399a8a) as a [URL](https://npiregistry.cms.hhs.gov/help-api/json-conversion) in the documentation for `npi_search()`.

> * The help page (or vignette?) for `npi_search()` could provide more details for how matching is done by default. For example, I get different results if I search for city = "new york" versus city = "new york city". I know you have not programmed the name matching yourself, but I think this is important information for the user to have so at least pointing to a link where this information can be found could be useful.

This is a great question, but I am unable to find an answer in the API documentation online.

> * It could be useful to add a message from `npi_search()` that explicitly states something along the lines of "Returning 10 of xxx records for the specified search parameters" so that the user is aware they are getting only a subset of the records that match their search parameters.

Great suggestion. When the user executes `npi_search()` the package now [produces](https://github.com/frankfarach/npi/commit/25553c2c7f6df1cb401010e3deb1859c32d9aef7) a message of the form "X records requested", where X is the number of records the user requested.

> * Why is 1200 the maximum for the `limit` argument in `npi_search()`?

That's the limit set by the API. In a section called "Note on NPI Downloadable Files", the advanced vignette [indicates](https://github.com/frankfarach/npi/commit/e656dc23acb02ed7534f89f76527d79f305c5006) how users can obtain the full dataset as a download from NPPES:

"CMS regularly releases full NPI data files [here](https://download.cms.gov/nppes/NPI_Files.html). We recommend that users download the data file if they need to work with the entire dataset. The API and `npi_search()` returns a maximum of 1,200 records. Also consider downloading the entire data if you need to work with more than the maximum. Data dissemination files are zipped and will exceed 4GB upon decompression. Additional documentation is provided [here](https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/DataDissemination)."

> #### Tests
> * Look thorough to me.

Thanks!

> #### Functionality/performance
> * No performance issues or areas for improvement that I can see
> * Your error messages are quite good, providing useful information about where errors in supplied arguments occurred and what is allowed. This is really user friendly!

Thank you. I learned a lot trying to incorporate best practices for error messages.

> * It seems like when invalid arguments are passed to `npi_search()` it still tries to query the API first and then returns an error. Internal checks could be done in this function to avoid the query altogether in these cases and to more quickly return a useful error message.

The package implements error-checking logic for which the API documentation is sufficiently clear. For example, since receiving reviewers' comments, the package now [checks for illegal characters](https://github.com/frankfarach/npi/commit/344ae6177df6057909fe9bd8abaf95de28fe9266) as those rules are clearly spelled out in the API documentation. The checks are performed for applicable user-supplied query parameters (i.e., `first_name`, `last_name`, `city`, `organization_name`) before sending the request to the API. Several unit tests have been added as well. It now also checks for correct usage of wildcard characters (see my response to @Rekyt). All of this new functionality is covered by unit tests.

> * I see that the taxonomies field in the result of `npi_search()` has a "primary" column, which I assume is used by `npi_summarize()` to identify the primary taxonomy. However, how does `npi_summarize()` determine the "primary" address and phone number when there are more than one listed, since there is no corresponding "primary" column in addresses? Is it always just the first record from taxonomies and address?

The phone number is the "Provider Business Practice Location Address Telephone Number" (field #35 [here](https://npiregistry.cms.hhs.gov/help-api/json-conversion), where the `address_purpose` field returned by the API is "LOCATION"). There is only one row for this field, and the it is simply called "phone" in the returned tibble.

`primary_practice_address` is assembled from individual address components corresponding to the "Primary Business Practice Location Address" (where the `address_purpose` field returned by the API is "LOCATION"). There is also only one row of these.

`primary_taxonomy` is the first row where `taxonomies_primary == TRUE`. In my experience, there is typically only one row per NPI record where this value is TRUE. Sometimes, however, there is only one row and the value is FALSE. The `npi` package will return the taxonomy for that row anyway under the premise that a single taxonomy is a primary taxonomy for practical summarization purposes.

> * Does `npi_flatten()` always return a sensible result? It seems like it merges together a lot of different pieces of information from different list columns but does it always make sense to cross-tabulate every single one? I don't have an idea of where it might go wrong, but always wonder when merging tibbles with different numbers of rows.

I appreciate your raising this excellent point.

`npi_flatten()` rigorously checks the inputs before manipulating the data, effectively implementing a data contract on behalf of the user:
* The input data frame must have a pre-determined S3 class which is applied only to results from `npi_search()`.
* The input data frame must have a specific structure, including correctly typed, named, and ordered columns.
* The checks are fully covered by unit tests.

However, despite [adding](https://github.com/frankfarach/npi/commit/365389993e8b84d6c3154dfe801064854187a235) several inbound data quality checks and associated unit tests, it's possible to use `npi_flatten()` to produce technically correct, but analytically impractical, results. Just as you indicated, the risk of this outcome grows in proportion to the number of list columns passed into the function. Depending on the user's goal, it may be fine to join everything in one go; in more complex situations, however, the more conservative approach would be to flatten one list-column at a time (e.g.,, npi-by-basic, npi-by-taxonomy, npi-by-identifier, etc.), transform each result as needed, and join them by `npi` in an order that yields a sensible result. Either way, `npi_flatten()` can help.

To highlight these nuances, the README now [indicates](https://github.com/frankfarach/npi/commit/0565f4dfc76933671c34dc59e656d7646db13a44) that users should explore the data carefully first to determine the best approach to assembling their target dataset. Beyond explaining what the function does and offering general guidance, I think it would be difficult to engineer guardrails that prevent these more nuanced problems up front -- though I am open to suggestions. Much like performing joins and other data manipulations with SQL, it's important to understand the data before extensively transforming it.
