### 2022-11-03 #505 @frankfarach
👋 @maelle - At long last, both of my responses are in! :) Thanks again for the excellent reviews and incredible patience while waiting for my responses. I look forward to answering any additional questions you, @zabore, or @Rekyt may have.
