### 2022-11-03 #505 @maelle
Thank you @frankfarach!!

@zabore & @Rekyt could you please confirm whether the responses are satisfying to you? Please use the [approval template](https://devguide.ropensci.org/approval2template.html) (when you're ready to "approve", that is). Thanks again for your reviews.
