### 2022-11-07 #505 @zabore
## Reviewer Response

Thanks for the thorough responses @frankfarach! I have nothing further to add.

#### Final approval (post-review)

Estimated hours spent reviewing: 4.5
