### 2022-11-08 #505 @Rekyt
## Reviewer Response

Thank you @frankfarach for all the very detailed responses! `npi` is very robust and well documented, I'm sure the interested community is going to love it.

#### Final approval (post-review)

Estimated hours spent reviewing: 3.5
