### 2022-11-08 #505 @ropensci-review-bot
Approved! Thanks @frankfarach for submitting and @Rekyt, @zabore for your reviews! :grin:
