### 2022-11-08 #505 @frankfarach
Thanks, @maelle, @zabore, and @Rekyt for collaborating on the peer review of the npi package! I believe it has improved the quality and usability of the package.

@maelle, I will work through the to-do list for transferring this package to rOpenSci.

Thanks again!
